+++ services/elasticsearch/struct_synonyms_dicts.go
@@ -18,7 +18,7 @@ package elasticsearch
 // SynonymsDicts is a nested struct in elasticsearch response
 type SynonymsDicts struct {
 	Type       string `json:"type" xml:"type"`
-	FileSize   int64  `json:"fileSize" xml:"fileSize"`
+	FileSize   int    `json:"fileSize" xml:"fileSize"`
 	SourceType string `json:"sourceType" xml:"sourceType"`
 	Name       string `json:"name" xml:"name"`
 }
